Replace hand-rolled ADXL345 I2C code with ADXL345_WE library
Our custom I2C implementation had bugs (wrong DATA_FORMAT register value,
undefined evaluation order in Wire.read() calls, no error checking).
Switched to wollewald/ADXL345_WE library for proper sensor handling.

Changes:
- Add ADXL345_WE library dependency to platformio.ini
- Refactor Accelerometer class to wrap library, return xyzFloat directly
- Change accel_raw_t typedef from int16_t to float (end-to-end floats)
- Update AccelSample/AccelSampleMsg to use floats (20 bytes per sample)
- Reduce batch size from 16 to 12 samples (calculated from ESP-NOW limit)
- Replace magic numbers with constants (ESP_NOW_MAX_PAYLOAD,
  ACCEL_MSG_HEADER_SIZE, ACCEL_SAMPLES_MAX_BATCH)
- Add static_assert to verify message fits in ESP-NOW payload
- Update telemetry CSV output format for float values

+++ shared/types.h
@@ -19,9 +19,9 @@ typedef uint64_t interval_t;
 // 32-bit is sufficient: max useful period ~1 second at 60 RPM = 1,000,000 µs
 typedef uint32_t period_t;
 
-// Accelerometer raw axis value (ADXL345: 256 LSB/g in full resolution)
+// Accelerometer axis value (float for compatibility with ADXL345_WE library)
 // Typedef'd so sensor changes only require updating this one place
-typedef int16_t accel_raw_t;
+typedef float accel_raw_t;
 
 // Time conversion macro
 #define SECONDS_TO_MICROS(s) ((s) * 1000000ULL)
